requirements_new.txt
# Основные зависимости
pandas>=1.5.0
numpy>=1.21.0
ta>=0.10.0
stable-baselines3>=2.0.0
gymnasium>=0.28.0
python-dotenv>=0.19.0
pybit>=5.0.0
scikit-learn>=1.1.0
joblib>=1.2.0

# Дополнительные зависимости для мониторинга
psutil>=5.9.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Зависимости для конфигурации
PyYAML>=6.0

# Зависимости для тестирования
pytest>=7.0.0
pytest-cov>=4.0.0

# Зависимости для разработки
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# Опциональные зависимости
optuna>=3.0.0  # Для оптимизации гиперпараметров
tensorboard>=2.10.0  # Для мониторинга обучения
